passed unit tests (put all matlab tests in one file, currently disabled)
